Split out constraint/parse.rs from constraint.rs
